Invoke PropertyChanged handler from UI thread
This was initially present, has been removed some time back, wondering why. In the absence of this wrapping, sometimes we get the "The application called an interface that was marshalled for a different thread" exception. Initially in the Template 10 project (when it was only Template10), it was present as `(App.BootStrapper).Dispatch(() =>{});`. In the current version I have referenced the path to the `DispatcherWrapper`. Maybe we can bring a reference into BootStrapper itself?