initial keyboard navigation for HamburgerMenu closes Windows-XAML/Template10#806
this is a solution to make keyboard controls more usable in the hamburger menu.  It largely does so through changing the container to a GridView rather than itemview where selection controls are more natural.   Its not perfect as we have to pass the selection down to the child button when the selected item changes.  When you only have one item this is problematic as its not transferred directly (so double tab stops occurs on things like settings when the item count is only 1).  The benefit is it is far more keyboard friendly now.